Add an ability to drag and drop items
New Features:
1. Added an ability to drag fitted items around: swap them or move an item to an empty slot. Added a special handling for fitted items with charges.

2. Added an ability to drag an item from Hardware to an individual fit slot. A support for dragging charges from the charges tab was not added in this pull-request.

Drag and drop feature is powered by a native [HTML Drag and Drop API](https://developer.mozilla.org/en-US/docs/Web/API/HTML_Drag_and_Drop_API).

There is a problem with the drag item image in case if user drags an item from the Hardware list for the first time. HTML Drag and Drop API supports a way to set an image like this:

```js
 let img = new Image();
 img.src = `https://images.evetech.net/types/${typeId}/icon?size=64`;
 e.dataTransfer.setDragImage(img, 32, 32);
```

Evetech images are not preloaded, and in case if Browser doesn't have the image cached/preloaded even if an image is set like in the code snipped above it wont be displayed, but it will on every other attempt to drag the same item.i