refactor: reworked all providers, so they are more well defined in their actions
This basically touches every file in this repository; so while at it, also
reworked how the storybooks work (you can now select different fits to test
with), and did a bunch of cleanup where I saw it.

There should be no functional difference, just mostly providers (and hooks)
that are named differently.